VDPipeline: Make aborted flushes wait for pending frames.
> Problem:

TL;DR: The VideoDecoderPipeline seems to be violating a requirement of
the media::VideoDecoder interface: when a flush is aborted, there might
still be frames in the pipeline (e.g the image processor) and under
certain circumstances, the VideoDecoderPipeline may output those frames
after calling the flush callback. The media::VideoDecoder interface
requires that the flush callback be called *after* all pending frames in
the queue are output.

This problem was detected with OOP-VD, but it's not unique to it. My
detailed analyisis is at b/261096749#comment3.

> Solution:

This CL fixes this by expanding the logic in [1]. Before this CL, that
logic made it so that we held onto the client flush callback only when
the underlying decoder told us that the flush was completed successfully
at that level. With this CL, we hold onto that callback regardless of
the status. That way, even if the underlying status tells us that the
flush is aborted, we won't call the client flush callback immediately
and instead we'll wait for pending frames to either be output or
discarded.

Note that there's a subtlety with this solution: we need to determine
what DecoderStatus to call the client flush callback with. In the
typical case, we just pass the one given to us by the underlying
decoder. However, there are cases in which we may want to pass a
different DecoderStatus. One such case is described as a comment in the
code.

[1] https://source.chromium.org/chromium/chromium/src/+/main:media/gpu/chromeos/video_decoder_pipeline.cc;l=612-616;drc=63c65d4c03ee4334e1202e2f4016cb9674fdab75

Bug: b:261096749
Test: video.Seek.h264_oopvd on asurada
Test: video.Seek.switch_h264_oopvd on asurada
Change-Id: Id2e98d4fbba601f3ec181be8ddcd26ec51817b3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4073952
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1087630}