VDPipeline: Abort pending frames on an error.
This CL is a follow-up to CL:4073952. That CL ensured that flush
callbacks are not called immediately when a flush request is aborted,
and instead, we wait until there are no more pending frames before
calling the flush callback.

The flush callback needs a DecoderStatus, so that CL saves the status
that it gets from the underlying VideoDecoderMixin when the flush is
completed at that level. That way, it can use that status at the moment
the flush callback is called. However, it's possible to override that
status by passing one to CallFlushCbIfNeeded(). There are two paths that
do this: OnResetDone() and OnError(). If CallFlushCbIfNeeded() gets
called from OnResetDone() or OnError() and there are pending frames in
the pipeline, then the "overriding" DecoderStatus will get dropped
because CallFlushCbIfNeeded() would early out. Fortunately,
OnResetDone() aborts pending frames before calling
CallFlushCbIfNeeded(), so in that path, the HasPendingFrames() call
inside CallFlushCbIfNeeded() should return false. This CL does the same
for OnError() and adds a DCHECK() documenting this expected behavior.

Bug: b:261096749
Test: video.Seek.h264_oopvd on asurada
Test: video.Seek.switch_h264_oopvd on asurada
Change-Id: Ic5a9deb060a45da8f440d823079e195c4fa4e30a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4128592
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Pilar Molina Lopez <pmolinalopez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1087637}