saved desks: Launch proper browser type for apps
Currently when an app is saved as a lacros browser, upon a restart,
the desk conversion logic will manually assign the app id as a
chrome app id instead of lacros because the flag to determine
lacros availability is not propagated from browser to the low
level system in time.

Since we only care about the browser type when we are launching
the app, we can make sure to check for lacros during the launch
step instead of just during the conversion step from proto.

BUG=b:261775603
TEST=Manually tested and confirmed expected behavior. Added unittest for this behavior.

Change-Id: I9ffc35085fc0438222d0f747f24bed6cc26f10ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4112705
Reviewed-by: Nancy Wang <nancylingwang@chromium.org>
Commit-Queue: Matthew Zhu <zhumatthew@google.com>
Reviewed-by: Richard Chui <richui@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1087647}