Update callers to check return value of new TestNavigationManager::Wait*
This is a no-op change to /content/browser/service_worker to update all callers of WaitForFirstYieldAfterDidStartNavigation and WaitForNavigationFinished to check for success. Failure could be due to a test timeout or other error in which case we should stop the test. Proceeding further with the test will just produce confusing output.

The return values will be marked `[[nodiscard]]` in a follow-up change.

This CL was uploaded by git cl split.

R=nhiroki@chromium.org

Bug: 1403181
Change-Id: I73f7b6d20a121e1937584190935171713592e732
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4122020
Auto-Submit: Fergal Daly <fergal@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1087702}