volume_manager: use FindExternalMountPoint to register FuseBox ADP
Based on crrev.com/1086295 for FSP volumes.

The DCHECK in the FuseBox ADP fsid register code fires if that fsid
is already registered with chrome storage. So remove the Profile if
test (irrelevant) and replace it with a FindExternalMountPoint test
to check the fsid is not already registered. If that is so, then go
ahead and register the FuseBox ADP fsid with chrome::storage.

Like crrev.com/1086295, use `fusebox_fsid` when registering the ADP
with chrome::storage, and when revoking that ADP.

Bug: b/263194149
Test: No change in behavior, covered by manual QA ADP tests.
Change-Id: Iba39fd7f19045bfd913834b725bd144475dcc884
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4128265
Reviewed-by: Ben Reich <benreich@chromium.org>
Commit-Queue: Noel Gordon <noel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1087715}