CCA: Resolve GIF recording when the preview stream is ended
Previously, we only use video.requestVideoFrameCallback() to record a
GIF video and resolve the promise when the recording is ended. However,
there is an issue when users record with an external camera, and they
unplug the cam while recording. Since the callback is never called again
and there's no other ways to resolve the promise.

To solve this issue, We add an event listener to listen when the video
stream ends and resolve the promise to end the GIF recording.

Bug: b:257397808
Test: Manually, tast run <DUT> camera.CCAUIRecordVideo camera.CCAUIRecordVideo.multi_stream
Change-Id: I1d6591828779b65a1a31300963b1b74468463e35
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4127739
Reviewed-by: Chu-Hsuan Yang <chuhsuan@chromium.org>
Commit-Queue: Kam Kwankajornkiet <kamchonlathorn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1087758}