[CodeHealth] Triage use of UnsafeDanglingUntriaged ptr in ProgressIndicatorAnimationDelegate
This CL is part of a larger cleanup effort to triage dangling pointers
and ensure they aren't the cause of memory issues in Chrome. See:
https://docs.google.com/document/d/164SnySRX6MZgFkX8qLLLClTnZ0S8s5wLiNFWNyQ8f4E/edit.

This change simply replaces UnsafeDanglingUntriaged with UnsafeDangling
with the conclusion that the lifetime of all the corresponding objects
extends beyond that of the delegate.

Bug: 1381411
Change-Id: Id43412c3468e40fd1496539b96a4e54472978cba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4111659
Reviewed-by: David Black <dmblack@google.com>
Commit-Queue: Ankush Singh <ankushkush@google.com>
Cr-Commit-Position: refs/heads/main@{#1087972}