[sync] Add histograms for addTrustedSyncEncryptionRecoveryMethod() calls
The goal is to understand the behavior around Javascript API
chrome.addTrustedSyncEncryptionRecoveryMethod(): the proportion of
invocations that fail and why.

Because the logic is plumbed through multiple layers, separate metrics
are introduced for the most relevant pieces.

Some of the metrics are instrumented for Android only, which is the
platform that requires most investigation. Instrumenting these metrics
for other platforms requires more intrusive changes and would anyway
lead to buckets with different semantics.

Change-Id: Ia524e1b6b06a3e8887cd9bd31b5960b01306af7d
Fixed: 1403195
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4120275
Commit-Queue: Rushan Suleymanov <rushans@google.com>
Reviewed-by: Maksim Moskvitin <mmoskvitin@google.com>
Auto-Submit: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Rushan Suleymanov <rushans@google.com>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1087976}