[sync] Add ClearMetadataWhileStopped method for stopped data types
Currently, OnSyncStopping is not called for data types which are already
stopped. This prevents clearing of metadata for such types. The CL
introduces a new method ClearMetadataWhileStopped to be implemented by
controller delegates/processors to clear metadata specifically when the
data type is not running. Note the following:
1. ClearMetadataWhileStopped doesn't do anything yet. The delegates will
have to implement it individually.
2. This method, behind the feature flag, will be called by
ModelTypeController::Stop when the data type is in NOT_RUNNING or FAILED
state.
3. DataTypeController::Stop() will now also be called from
SyncServiceImpl when the sync engine is not running and/or from
ModelLoadManager when the data type is already stopped.

Bug: 897628
Change-Id: I159b2eb232fcfe19250338e77eab2e3e52740a11
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4047372
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Ankush Singh <ankushkush@google.com>
Cr-Commit-Position: refs/heads/main@{#1087979}