Chrome for Testing trybots: Enable NaCl
The CfT trybots should mirror vanilla Chromium as much as possible. Omit
the explicit `disable_nacl` config so that GN defaults end up enabling
NaCl.

The purpose of this CL is to fix a trybot failure in the CI, which
currently tries to run nacl_loader_unittests, which fails when NaCl is
disabled.

Bug: 1401477
Fixed: 1401477
Change-Id: I28972f234fa9399eaea56a0e4121b2fd10d9e9f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4111976
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>
Commit-Queue: Thiago Perrotta <tperrotta@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1088015}