[js-code-coverage] Exclude uninteresting lines from coverage report
Coverage is reported for empty lines, comments and imports such that if
the file is simply imported once, the line is "covered". This
unfortunately has implications on the line coverage metric as a file
with simply imports and a single function that doesn't get covered could
have some inflated line coverage like 10/15 (66%) lines covered which is
incorrect as the only meaningful function in there is not.

This work excludes empty lines, full line comments (e.g. // not /* */)
and any lines that indicate an import.

Bug: 1402996
Test: ./merge_js_lib_test.py
Change-Id: Ica42c132fa66028fcc379146290ba99c54bcddbf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4120383
Reviewed-by: Zhaoyang Li <zhaoyangli@chromium.org>
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>
Auto-Submit: Ben Reich <benreich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1088076}