[VCManager] Remove VcWebApp in RemoveMediaApp
Background: `VideoConferenceWebApp` observes `WebContentsDestroyed`
and `PrimaryPageChanged`. Both methods call `RemoveMediaApp` on
`VideoConferenceClientImpl`. However, while the former also results
in the destruction of the VcWebApp, the latter does not.

This was causing an invariant violation as `RemoveMediaApp` removes
the entry associated with the passed VcWebApp `id` from its
internal map but the associated `VideoConferenceWebApp` was not
destroyed.

This CL adds a check in `RemoveMediaApp` to ensure the VcWebApp
associated with the `id` being removed is also destroyed.

Bug: b/263332944
Test: Manually tested on brya DUT.
Change-Id: I4fcd6f26a8040f43d71e740fa8c23746af9c8147
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4127578
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Auto-Submit: Sanjeet Dasharath <sanjeetnd@google.com>
Commit-Queue: Sanjeet Dasharath <sanjeetnd@google.com>
Reviewed-by: Charles Zhao <charleszhao@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1088084}