[FilesRefresh] Add color tokens for shadow/pressed
The shadow color is from the spec, the pressed color is blending
the ripple color on top of the hover color.

Demo: http://go/scrcast/NTE5OTkxOTA0MTIxNjUxMnxlODI3YTAyNy0wZA

The icon button in the demo has a hover color and a paper ripple,
when it's on (e.g. the dropdown is open), we need to maintain
the same color as clicked, hence the introduction of pressed
color.

Bug: b:242938039
Change-Id: I4008f40fac4fb9c883c9b722309a81230bdf07a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4122014
Reviewed-by: Zain Afzal <zafzal@google.com>
Commit-Queue: Wenbo Jie <wenbojie@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1088102}