Remove icon effects and the icon type from the icon writer key.
This CL is used to prepare for CL:4122779.

Icon effects and the icon type are not used to get the raw icon image data. The app id, icon size and scale factor are used to get the raw icon image data.

If there is a icon data request, we don't need to call GetCompressedIconData to get the icon data again for the same app id,
icon size and scale factor. So modify `pending_results_` to use the app id, icon size and scale factor as the key to prevent calling duplicated GetCompressedIconData at the same time for the same app id, icon size, and scale factor.

Add unit tests to verify that we can get the correct icons with and
without icon effects using separate icon loadings, or calling icon
icon loadings parallel at the same time.

BUG=1380608

Change-Id: I801ea820c9db32de3149d0fedf085b22749f0875
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4125301
Reviewed-by: Tim Sergeant <tsergeant@chromium.org>
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1088110}