Update callers to check return value of new TestNavigationManager::Wait*
This is a no-op change to /chrome/browser/subresource_filter to update all callers of WaitForFirstYieldAfterDidStartNavigation and WaitForNavigationFinished to check for success. Failure could be due to a test timeout or other error in which case we should stop the test. Proceeding further with the test will just produce confusing output.

The return values will be marked `[[nodiscard]]` in a follow-up change.

This CL was uploaded by git cl split.

R=alexmt@chromium.org

Bug: 1403181
Change-Id: I159e6dbe96d805102cdadfc539350d7e4afa905a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4122376
Auto-Submit: Fergal Daly <fergal@chromium.org>
Reviewed-by: Alex Turner <alexmt@chromium.org>
Commit-Queue: Alex Turner <alexmt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1088117}