Add import of paper-tooltip.js and cr-icon-button.js
Adding this in seems to fix the issue of `this.$.tooltip.show() is not
defined`. This seems to be because the typescript deletes the import
line for the type, unless it is explicitly called with new. For polymer
elements, the code has to be run for it to work correctly, so we need to
explicitly add back the import.

This was deleted in https://crrev.com/c/3920251.

This also adds an import of cr-icon-button.

BUG=b:264211466

Change-Id: Icb38510cb0e294d8b1bf5b2a8f74ff0fbabe905a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4129578
Reviewed-by: Michael Cui <mlcui@google.com>
Reviewed-by: Mehrab N <mehrab@chromium.org>
Commit-Queue: Keith Lee <keithlee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1088123}