Privacy Sandbox Settings: Treat SESSION_ONLY as ALLOW in settings
For some cookie content settings, the SESSION_ONLY state is not
accurate and an ALLOW setting is more accurate. This CL changes the behavior of the site settings helper to alias SESSION_ONLY to ALLOW
when providing information to the front end.

It does this only when the Privacy Sandbox settings release 4 flag
is enabled, as this removes the location for affected settings to
be displayed.

CL also removes some spurious logic that inserts dummy "default"
parent exceptions into the returned exception list. This _appears_
to be a legacy holdover, as the only location which uses this logic
filters any default exceptions out. Removing it allows for more
precise testing of the returned exceptions, as they exactly match
what is in the Host Content Settings Map.


Bug: 1378703
Change-Id: I5cd2e92d20e93f757cd2b3329c126ea3907e7ff8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4129525
Reviewed-by: Olesia Marukhno <olesiamarukhno@google.com>
Commit-Queue: Theodore Olsauskas-Warren <sauski@google.com>
Cr-Commit-Position: refs/heads/main@{#1088153}