Update callers to check return value of new TestNavigationManager::Wait*
This is a no-op change to /chrome/browser/site_isolation to update all callers of WaitForFirstYieldAfterDidStartNavigation and WaitForNavigationFinished to check for success. Failure could be due to a test timeout or other error in which case we should stop the test. Proceeding further with the test will just produce confusing output.

The return values will be marked `[[nodiscard]]` in a follow-up change.

This CL was uploaded by git cl split.

R=wjmaclean@chromium.org

Bug: 1403181
Change-Id: I69d9ff13d385c8af29a9bd14ffbb68e5e978dd58
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4123610
Auto-Submit: Fergal Daly <fergal@chromium.org>
Commit-Queue: James Maclean <wjmaclean@chromium.org>
Reviewed-by: James Maclean <wjmaclean@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1088243}