capture-mode-demo-tools: Revise timer logic
As mentioned in b/262923676, there are inconsistencies when the
user intends to release all the keys. Different behaviors may
occur due to the key release order. This CL implements the design
in http://b/262923676#comment3 by introducing the timer with a
different delay to hold on the key combo viewer refresh until
the user's intention becomes clear i.e. update the key combo
viewer if the gap between two key up is larger than the threshold
and hide the entire key combo viewer with existing hide timer
otherwise. The timer with shorter duration will be triggered on
key up of any modifier key when there is no ongoing hiding process.

This CL also revises to timer logic so that the hide timer will be
triggered on key up of the the non-modifier key with no modifier keys
pressed or on key up of the last modifier key with no non-modifier
key that can be displayed independently pressed.

The demo for the CL is available at: http://b/262923676#comment4

Fixed: b:262923676
Test: Manually on DUT + added unit tests
Change-Id: Ice0d40ce0f57036c09c3203840961619b9a1198a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4119745
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Commit-Queue: Michele Fan <michelefan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1088246}