Remove RenderViewHostChanged from AshWebViewImpl
RenderViewHostChanged is an old API which is deprecated and we would
like to remove. This CL moves AshWebViewImpl to the newer
PrimaryPageChagned API. The existing frame based observer already
restricts itself to the outermost main frame, but the
RenderFrameHostChanged event is not fired for the very first
navigation a new WebContents. Therefore using the PrimaryPageChanged
is the more correct change in this case.

Bug: 1168562, 1394513
Change-Id: I0691c9a19d606d0710802191c5f41eb2b6d61d8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4092763
Auto-Submit: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1088282}