[omnibox] Cleans up some of omnibox UI class roles and meanings
Makes ChromeOmniboxClient receive a ChromeOmniboxEditController instance
rather than a base OmniboxEditController instance and downcasting it.

Makes ChromeOmniboxEditController an abstract class which is implemented
in tests and by LocationBarView. This class is intended as an abstract
class which is currently ensured via using a protected constructor.

Bug: 1404748
Change-Id: I2c49195d1f6d1dc3c58b93943273161a96ec33b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4130888
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1088322}