Update callers to check return value of new TestNavigationManager::Wait*
This is a no-op change to /chrome/browser/ssl to update all callers of WaitForFirstYieldAfterDidStartNavigation and WaitForNavigationFinished to check for success. Failure could be due to a test timeout or other error in which case we should stop the test. Proceeding further with the test will just produce confusing output.

The return values will be marked `[[nodiscard]]` in a follow-up change.

This CL was uploaded by git cl split.

R=carlosil@chromium.org

Bug: 1403181
Change-Id: Ib7b4c850715e7b11f0d0c6d97a48c1ff0f14ce10
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4123611
Auto-Submit: Fergal Daly <fergal@chromium.org>
Commit-Queue: Carlos IL <carlosil@chromium.org>
Reviewed-by: Carlos IL <carlosil@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1088328}