ax_ui: chromium hangs on user input when dealing with large documents
SnapToMaxTextOffsetIfBeyond can be a heavy operation when triggered
on ax position creation. Disable it for now since it regresses the user
experience harsh. Later, all callers should be reworked, the method call
removed and only the IsValid() assertion should be left in place.

This is a partial revert of crrev.com/c/3810575, and todo-marked tests
from crrev.com/c/3961717 which was landed on top of the original
culprit CL.

Regressed by: crrev.com/c/3810575

Bug: 1401591
Bug: 1396198
Change-Id: I448cdc4135a362f708882c0223a01507a6dbb7b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4126876
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1088332}