Pass a bool flag instead of optional<int> to SetUpFieldTrials
Instead of passing an optional<int> for low_entropy_provider_source to
just decide whether to record a variation id for it or not. This CL
tries to be more intentional about it and pass a bool flag to control
the addition of variation id and get the low_entropy_provider_source
from MetricsStateManager.

Bug: b/183955043
Bug: b/263797385
Bug: 1191479
Change-Id: I0919b97f63e57c79fec30197b0d8fac8ba4bfa7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4111929
Reviewed-by: Steven Holte <holte@chromium.org>
Auto-Submit: Hazem Ashmawy <hazems@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Commit-Queue: Peter Beverloo <peter@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1088367}