Fix an issue that duplicate download file cannot be opened
When showing the duplicate download dialog, chrome passes a file path
to check whether the file name exists in content resolver. However,
the file path cannot be used to open the file due to Android security
restrictions. Instead, it should used the content URI that has the
file name.
This CL fixes the issue that Chrome is using the file path when
opening the download link on duplicate download dialog.

BUG=1402971

Change-Id: I0d055be630da8ef9ff61f9c12c8e57c42c2a8cba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4119079
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: Min Qin <qinmin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1088395}