Update callers to check return value of new TestNavigationManager::Wait*
This is a no-op change to /chrome/browser/download to update all callers of WaitForFirstYieldAfterDidStartNavigation and WaitForNavigationFinished to check for success. Failure could be due to a test timeout or other error in which case we should stop the test. Proceeding further with the test will just produce confusing output.

The return values will be marked `[[nodiscard]]` in a follow-up change.

This CL was uploaded by git cl split.

R=dtrainor@chromium.org

Bug: 1403181
Change-Id: Ia174ffc583a413294f72935c81486872b1f47721
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4113718
Reviewed-by: David Trainor <dtrainor@chromium.org>
Auto-Submit: Fergal Daly <fergal@chromium.org>
Commit-Queue: David Trainor <dtrainor@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1088423}