Refactor Safe Browsing checks to run in individual objects
This CL refactors SafeBrowsingUrlCheckerImpl to move its mechanism-
specific logic to individual SafeBrowsingLookupMechanism objects. These
objects are responsible for starting the checks and reporting back the
results to the caller. Each run is wrapped by a runner object named
SafeBrowsingLookupMechanismRunner that is responsible for canceling the
checks when there is a timeout.

The two current Safe Browsing mechanisms are the URL-based real-time
lookup mechanism and the hash-based database lookup mechanism, and each
of these has its own class extending SafeBrowsingLookupMechanism for
mechanism-specific logic. Since the URL-based real-time lookup mechanism
can also sometimes default back to hash-based database lookups, it can
create its own hash-based database lookup mechanism object, await the
response, and then report back the results to the original caller.

This change is purely a refactor. The goal is to make it easier and
cleaner for the HPRT experiment to be able to run multiple mechanisms at
a time. However, it has the added benefit of reducing the responsibilities that SafeBrowsingUrlCheckerImpl is responsible for, so
this code should remain compartmentalized like this even after the
experiment is deprecated. This is especially relevant considering the
plans to add an additional mechanism.

Change-Id: Ie1c5a50ae9cc17ca9d5acdb5ae6e095515810dbe
Bug: 1392144
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4035695
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Commit-Queue: thefrog <thefrog@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1088444}