Fix condition for forcing incognito mode in chrome::NewEmptyWindow
The flag `browser_defaults::kAlwaysOpenIncognitoWindow` means
"Should always open incognito windows when started with --incognito
switch?"
as the comment above the declaration of the flag says. So the flag
should interfere only with `--incognito` switch.

But in the body of the function `chrome::NewEmptyWindow` before this
patch, if the flag was set to false (it is set to false unless in case
of ChromeOS), it suppressed not only `--incognito` switch but also
erroneously `IncognitoModePrefs::Availability::kForced` setting in
preferences.

This patch fixes the behavior so that the flag
`browser_defaults::kAlwaysOpenIncognitoWindow` only interferes with
`--incognito` switch and not with
`IncognitoModePrefs::Availability::kForced` setting in preferences.

After this fix of the behavior in the body of the function
`chrome::NewEmptyWindow`, `app_controller_mac` can pass
`profile->GetOriginalProfile()` to `chrome::NewEmptyWindow` safely
without worrying of incognito mode not being forced when it should.

Bug: 1404143
Fixed: 1371923
Change-Id: I248bed7a129d5df85195580b4b187b4e5d6d1b39
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4129574
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1088460}